+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## New Features
 
 * Show "total until now" in work log tab
+* Move custom filters to the general directory (#100)
 * Import/export data (#119)
 * Show next save/backup/synchronization date (#121)
 * Add line between rows in task table (#127)